Time optimizations, refactoring, modularizations
