+++ components/Activity.js
@@ -6,16 +6,21 @@ class Activity extends React.Component{
     super(props);
   }
 
+  navigate = () => {
+    this.props.navigation.navigate('Activity', {
+      title : this.props.exercise.title,
+      sets  : this.props.exercise.sets,
+      reps  : this.props.exercise.reps,
+      weight: this.props.exercise.weight,
+      notes : this.props.exercise.notes,
+    });
+    
+    
+  }
+
   render(){
     return(
-      <TouchableOpacity style={styles.activityContainer} onPress={() => this.props.navigation.navigate('Activity', {
-        title : this.props.exercise.title,
-        sets  : this.props.exercise.sets,
-        reps  : this.props.exercise.reps,
-        weight: this.props.exercise.weight,
-        notes : this.props.exercise.notes,
-        userId: this.props.userId
-      })}>
+      <TouchableOpacity style={styles.activityContainer} onPress={this.navigate}>
         <Text style={{color:'#ff7675', fontWeight:'bold'}}>{this.props.exercise.title}</Text>
       </TouchableOpacity>
     )
