--- components/StyledText.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native';
-
-export class MonoText extends React.Component {
-  render() {
-    return <Text {...this.props} style={[this.props.style, { fontFamily: 'space-mono' }]} />;
-  }
-}
