+++ package.json
@@ -12,13 +12,15 @@
   },
   "dependencies": {
     "@expo/samples": "2.1.1",
+    "@react-native-community/async-storage": "^1.5.0",
     "axios": "^0.18.0",
     "expo": "^32.0.0",
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
     "react-native-elements": "^1.1.0",
     "react-native-prompt-crossplatform": "rgomezp/react-native-prompt-crossplatform#custom-keyboard",
-    "react-navigation": "^3.0.9"
+    "react-navigation": "^3.0.9",
+    "why-did-you-update": "^1.0.6"
   },
   "devDependencies": {
     "babel-preset-expo": "^5.0.0",
